% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DependenciesFinder.R
\name{addLibraries}
\alias{addLibraries}
\title{Add libraries to Libraries list or initiazlite it}
\usage{
addLibraries(newLibraries, outputFileName, fileName = NULL)
}
\arguments{
\item{newLibraries:}{a list of new libraries the user wants to add to the
previous one.}

\item{outputFileName:}{name of the output csv file.}

\item{fileName:}{(optional) the path to a csv with libraries.}
}
\description{
Function reads a csvFile with almost 1 column named "Package"
 (if provided), and a vector of libraries. Then, it executes getDependencies
 function with both libraries lists joined, and save it on a new csv file. It
 also test the results for better reliability. The csv file returned has 2
 columns: numeric id and "Package".
}
\author{
Italo Garleni
}
