% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DependenciesFinder.R
\name{addVersion}
\alias{addVersion}
\title{Add version to Libraries list}
\usage{
addVersion(fileName, outputFilename)
}
\arguments{
\item{fileName:}{the path to the library csv.}

\item{outputFileName:}{name of the output csv file.}
}
\description{
Function reads a csvFile with almost 1 column named "Package".
 Then, it checks libraries
version on local machine and save it on a new csv file. The csv file
returned has 3 columns: numeric id, "Package", and "Version".
}
\author{
Italo Garleni
}
