% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DependenciesFinder.R
\name{getDependencies}
\alias{getDependencies}
\title{Search for libraries dependencies}
\usage{
getDependencies(libraries)
}
\arguments{
\item{libraries:}{a list of libraries the user wants to analyze.}
}
\value{
vector with dependencies and library, sorted by its dependency.
}
\description{
Function that returns libraries' dependencies over a list of
 libraries, sorted by its dependency. Being ['n'=length(libraries)] and
 [1<='i'<='n'], library 'i' depends (or not) on the previous '1' to 'i'
 libraries, but never the other way around.
}
\author{
Italo Garleni
}
